backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlmodel==0.0.24
SQLAlchemy==2.0.39
alembic==1.15.2
pymysql==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic-settings==2.8.1
pyotp==2.9.0
webauthn==2.5.2
email-validator==2.2.0
httpx==0.28.1
